DAT-1118 refactor import classes
