+++ cmdb/check_routine.py
@@ -32,7 +32,10 @@ from cmdb.models.user_model.user import UserModel
 from cmdb.models.location_model.cmdb_location import CmdbLocation
 from cmdb.models.section_template_model.cmdb_section_template import CmdbSectionTemplate
 from cmdb.models.reports_model.cmdb_report_category import CmdbReportCategory
-from cmdb.user_management.constants import __FIXED_GROUPS__, __COLLECTIONS__ as USER_MANAGEMENT_COLLECTION
+from cmdb.models.user_management_constants import (
+    __FIXED_GROUPS__,
+    __COLLECTIONS__ as USER_MANAGEMENT_COLLECTION
+)
 from cmdb.framework.constants import __COLLECTIONS__ as FRAMEWORK_CLASSES
 
 from cmdb.errors.system_config import SectionError
