+++ cmdb/importer/mapper/__init__.py
@@ -13,18 +13,3 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
-"""TODO: document"""
-from cmdb.interface.rest_api.responses.messages.import_message import ImportMessage
-# -------------------------------------------------------------------------------------------------------------------- #
-
-class ImportFailedMessage(ImportMessage):
-    """Message wrapper for failed imported objects"""
-
-    def __init__(self, error_message: str, obj: dict = None):
-        """Init message
-        Args:
-            error_message: reason why it failed - exception error or something
-            obj (optional): failed dict
-        """
-        self.error_message = error_message
-        super().__init__(obj=obj)
