+++ cmdb/interface/rest_api/responses/__init__.py
@@ -24,7 +24,6 @@ from .insert_single_response import InsertSingleResponse
 from .update_single_response import UpdateSingleResponse
 from .update_multi_response import UpdateMultiResponse
 from .login_response import LoginResponse
-from .importer_object_response import ImporterObjectResponse
 from .gridfs_response import GridFsResponse
 # -------------------------------------------------------------------------------------------------------------------- #
 
@@ -39,6 +38,5 @@ __all__ = [
     'UpdateSingleResponse',
     'UpdateMultiResponse',
     'LoginResponse',
-    'ImporterObjectResponse',
     'GridFsResponse',
 ]
