+++ cmdb/interface/rest_api/routes/importer_routes/importer_route_utils.py
@@ -25,7 +25,7 @@ from werkzeug.wrappers import Request
 from cmdb.manager.query_builder.builder_parameters import BuilderParameters
 from cmdb.manager.types_manager import TypesManager
 
-from cmdb.importer.importer_helper import load_parser_class
+from cmdb.importer.helper.importer_helper import load_parser_class
 from cmdb.models.user_model.user import UserModel
 from cmdb.models.type_model.type import TypeModel
 from cmdb.security.acl.permission import AccessControlPermission
